pyproject.toml
[project]
name = "onshape2xacro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "onshape-robotics-toolkit",
    "tyro",
    "lxml",
    "pyyaml",
    "xacro>=2.1.1",
    "keyring",
    "cadquery-ocp>=7.7.0",
    "trimesh>=4.11.1",
    "scipy>=1.17.0",
    "pymeshlab>=2025.7",
    "cadquery>=2.6.1",
]

[project.scripts]
onshape2xacro = "onshape2xacro.cli:main"
validate-xacro = "onshape2xacro.validate:main"
visualize-export = "onshape2xacro.visualize_export:main"
simulate-robot = "onshape2xacro.simulate_robot:main"
fix-xacro = "onshape2xacro.fix_xacro:main"
xacro-to-urdf = "onshape2xacro.xacro_to_urdf:main"

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest",
]
visualize = [
    "matplotlib",
]
simulate = [
    "pybullet",
    "numpy",
]
